Create README.md
This little program is just a project I made to play with different parts of Python. Not to mix up I made 3 files that are connected to each other but to use you should simply personalize them. First you need to personalize "myappdb" with your user and host to be able to create a data base of your own. Then in "Myfirstapp", you'll see a structure of little form or app that I created with 2 users that I made a file to save their username & pass. Also I connected my forms to my data base and to make some fun I had put a little game in it for my sec. user to roll a dice before accessing to his form. and in "import subprocess", you'll make an exe file for your users, so they won't need to go through vscode to run the form!
after all it's just a play specially with different abillities of tkinter library and also getting to know some other good libraries. Fell free to play with my codes and personalize them for yourself to learn better:)